+++ src/components/Checkbox/Checkbox.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { ComponentMeta, ComponentStory } from '@storybook/react'
-import Checkbox from './Checkbox'
+import { Checkbox } from './Checkbox'
 
 export default {
   title: 'Atom/CheckBox',
