+++ src/components/Picker/Picker.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { ComponentMeta, ComponentStory } from '@storybook/react'
-import Picker from './Picker'
+import { Picker } from './Picker'
 
 export default {
   title: 'Atom/Picker',
