+++ src/styles/foundation/colors/colors-type.ts
